[form-factors] Move process traits to separate header file
